订单评价

+++ sky-server/src/main/java/com/sky/service/PinglunService.java
@@ -8,7 +8,7 @@ import java.util.List;
 // ===== service/PinglunService.java =====
 public interface PinglunService {
     void publishComment(String details, Integer shequid, Long userId);
-    void like(Integer id);
+    void like(Long id);
     void delete(Integer id);
 
     List<PinglunDetailDTO> getPinglunByShequId(Integer shequId);
